Split preprocess into different files.
